Extract QueryCriterion to interface
